protocol: increase HTLC precision to 1/1000 satoshi.
This gets truncated for on-chain transactions (thus, rounding may
contribute to fees).

This also means we currently have an upper bound of 0.04 BTC per HTLC;
this can be increased later if required.

+++ pkt.h
@@ -100,7 +100,7 @@ struct pkt *update_pkt(const tal_t *ctx,
  */
 struct pkt *update_htlc_add_pkt(const tal_t *ctx,
 				const struct sha256 *revocation_hash,
-				u64 value,
+				u32 value,
 				const struct sha256 *htlc_rhash,
 				u32 abs_locktime_seconds);
